+++ src/GitHub_Updater/Rest_Update.php
@@ -57,7 +57,7 @@ class Rest_Update extends Base {
 		$plugin           = null;
 		$is_plugin_active = false;
 
-		foreach ( (array) Plugin::instance()->get_plugin_configs() as $config_entry ) {
+		foreach ( (array) Class_Factory::get_instance( 'Plugin' )->get_plugin_configs() as $config_entry ) {
 			if ( $config_entry->repo === $plugin_slug ) {
 				$plugin = $config_entry;
 				break;
@@ -108,7 +108,7 @@ class Rest_Update extends Base {
 	public function update_theme( $theme_slug, $tag = 'master' ) {
 		$theme = null;
 
-		foreach ( (array) Theme::instance()->get_theme_configs() as $config_entry ) {
+		foreach ( (array) Class_Factory::get_instance( 'Theme' )->get_theme_configs() as $config_entry ) {
 			if ( $config_entry->repo === $theme_slug ) {
 				$theme = $config_entry;
 				break;
@@ -351,7 +351,7 @@ class Rest_Update extends Base {
 	 * @return bool|array $response
 	 */
 	private function parse_bitbucket_webhook( $request_body ) {
-		Basic_Auth_Loader::instance( parent::$options )->load_authentication_hooks();
+		Class_Factory::get_instance( 'Basic_Auth_Loader', parent::$options )->load_authentication_hooks();
 
 		$request_data = json_decode( $request_body, true );
 
@@ -364,7 +364,7 @@ class Rest_Update extends Base {
 
 		//$response['payload'] = $new;
 
-		Basic_Auth_Loader::instance( self::$options )->remove_authentication_hooks();
+		Class_Factory::get_instance( 'Basic_Auth_Loader', parent::$options )->remove_authentication_hooks();
 
 		return $response;
 	}